Test improved.
