+++ tests/Feature/Stories/GroupAdministratorCanUpdateGroupPhotoTest.php
@@ -1,0 +1,90 @@
+<?php
+
+
+namespace Tests\Feature\Stories;
+
+
+use App\Group;
+use App\User;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Storage;
+use Tests\TestCase;
+
+class GroupAdministratorCanUpdateGroupPhotoTest extends TestCase
+{
+    /**
+     * @test testUserCanUpdateAnOwnedGroupWithPhoto
+     */
+    final public function testUserCanUpdateAnOwnedGroupWithPhoto()
+    {
+        Storage::fake('gcs');
+        $user = factory(User::class)->create();
+        $userB = factory(User::class)->create();
+        $group = factory(Group::class)->create();
+        $group->user_uuid = $userB->uuid;
+        $group->save();
+        $user->groups()->attach($group, [
+            'is_admin' => true,
+            'status' => Group::ACCEPTED_STATUS
+        ]);
+        $response = $this->actingAs($user, 'api')->json('PUT', '/api/groups/' . $group->uuid, [
+            'photo' => UploadedFile::fake()->image('photo50.jpg'),
+            'name' => 'New name',
+            'is_public' => true,
+            'coordinates' => '10.1,50.5'
+        ]);
+        $response->assertOk();
+        $content = $response->getContent();
+        Storage::disk('gcs')->assertExists('photos/' . (json_decode($content))->photo->photo_url);
+
+        $response->assertJsonStructure([
+            'uuid',
+            'user_uuid',
+            'photo_uuid',
+            'name',
+            'photo_url',
+            'created_at',
+            'updated_at',
+            'user' => [
+                'uuid',
+                'user_uuid',
+                'photo_uuid',
+                'name',
+                'email',
+                'email_verified_at',
+                'free',
+                'photo_url',
+                'location' => [
+                    'type',
+                    'coordinates',
+                ],
+                'is_public',
+                'created_at',
+                'updated_at',
+                'love_reactant_id',
+                'love_reacter_id',
+                'is_admin',
+            ] ,
+            'photo' => [
+                'location' => [
+                    'type',
+                    'coordinates',
+                ],
+                'uuid',
+                'user_uuid',
+                'comment_uuid',
+                'ext',
+                'photo_url',
+                'is_public',
+                'created_at',
+                'updated_at',
+                'love_reactant_id',
+            ]
+        ]);
+        $response->assertJsonFragment([
+            'uuid' => $group->uuid,
+            'name' => 'New name',
+        ]);
+
+    }
+}
